Fisher's test on number of shared and unique intervals.

Fisher's test on number of shared and unique intervals.

bed_fisher(x, y, genome)

interval_union(x)

Arguments

x

tbl of intervals

y

tbl of intervals

genome

tbl of chrom sizes

Value

data_frame

See also

http://bedtools.readthedocs.org/en/latest/content/tools/fisher.html

Other interval-stats: bed_absdist, bed_jaccard, bed_reldist

Examples

x <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 10,     20,
  "chr1", 30,     40,
  "chr1", 51,     52
)

y <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 15,     25,
  "chr1", 51,     52
)

genome <- tibble::tribble(
  ~chrom, ~size,
  "chr1", 500
)

bed_fisher(x, y, genome)

#>   estimate     p.value conf.low conf.high                             method
#> 1      Inf 0.003846154 2.906919       Inf Fisher's Exact Test for Count Data
#>   alternative
#> 1   two.sided
#>