convienence function for generating example bed glyphs

convienence function for generating example bed glyphs

bed_glyph(x, y = NULL, .fun, ..., x_name = "X", y_name = "Y",
  start_col = "start", end_col = "end", title_ops = "(x, y)")

Arguments

x

tbl of intervals passed to .fun

y

optional tbl of intervals to pass to .fun (default = FALSE)

.fun

valr function to execute on (x and/or y intervals)

...

other arguments to pass to .fun (x ivls and y ivls passed by default if supplied)

x_name

name of plotted x intervals (default is "X")

y_name

name of plotted y intervals (default is "Y")

start_col

name of column containing start positions to use for plotting results (default = "start")

end_col

name of columm containing end positions to use for plotting results (default = "end")

title_ops

options to pass to title display in ggplot object (default = "(x, y"))

Value

ggplot2_obj