Divide intervals into new sub-intervals ("windows").

Divide intervals into new sub-intervals ("windows").

bed_makewindows(x, genome, win_size = 0, step_size = 0, num_win = 0,
  reverse = FALSE)

split_interval(interval, genome, win_size, step_size, num_win, reverse)

Arguments

x

tbl of intervals

genome

genome file with chromosome sizes

win_size

divide intervals into fixed-size windows

step_size

size to step before next window

num_win

divide intervals to fixed number of windows

reverse

reverse window numbers

interval

row of data frame

Value

data_frame with win_id column that contains a numeric identifier for the window.

Functions

split_interval: Helper to divide interval into labeled sub-intervals.

Note

The name and win_id columns can be used to create new interval names (see 'namenum' example below) or in subsequent group_by operations (see vignette).

See also

Other utils: bed12_to_exons

Examples

genome <- tibble::tribble(
 ~chrom, ~size,
 "chr1", 5000,
 "chr2", 400
)

x <- tibble::tribble(
  ~chrom, ~start, ~end, ~name, ~score, ~strand,
  "chr1", 100,    200,  'A',   '.',    '+',
  "chr2", 300,    350,  'B',   '.',    '-'
)

# Fixed number of windows 
bed_makewindows(x, genome, num_win = 10)

#> # A tibble: 20 × 7
#>    chrom start   end  name score strand win_id
#>    <chr> <dbl> <dbl> <chr> <chr>  <chr>  <dbl>
#> 1   chr1   100   110     A     .      +      1
#> 2   chr1   110   120     A     .      +      2
#> 3   chr1   120   130     A     .      +      3
#> 4   chr1   130   140     A     .      +      4
#> 5   chr1   140   150     A     .      +      5
#> 6   chr1   150   160     A     .      +      6
#> 7   chr1   160   170     A     .      +      7
#> 8   chr1   170   180     A     .      +      8
#> 9   chr1   180   190     A     .      +      9
#> 10  chr1   190   200     A     .      +     10
#> 11  chr2   300   305     B     .      -      1
#> 12  chr2   305   310     B     .      -      2
#> 13  chr2   310   315     B     .      -      3
#> 14  chr2   315   320     B     .      -      4
#> 15  chr2   320   325     B     .      -      5
#> 16  chr2   325   330     B     .      -      6
#> 17  chr2   330   335     B     .      -      7
#> 18  chr2   335   340     B     .      -      8
#> 19  chr2   340   345     B     .      -      9
#> 20  chr2   345   350     B     .      -     10
#> 

# Fixed window size
bed_makewindows(x, genome, win_size = 10)

#> # A tibble: 15 × 7
#>    chrom start   end  name score strand win_id
#>    <chr> <dbl> <dbl> <chr> <chr>  <chr>  <dbl>
#> 1   chr1   100   110     A     .      +      1
#> 2   chr1   110   120     A     .      +      2
#> 3   chr1   120   130     A     .      +      3
#> 4   chr1   130   140     A     .      +      4
#> 5   chr1   140   150     A     .      +      5
#> 6   chr1   150   160     A     .      +      6
#> 7   chr1   160   170     A     .      +      7
#> 8   chr1   170   180     A     .      +      8
#> 9   chr1   180   190     A     .      +      9
#> 10  chr1   190   200     A     .      +     10
#> 11  chr2   300   310     B     .      -      1
#> 12  chr2   310   320     B     .      -      2
#> 13  chr2   320   330     B     .      -      3
#> 14  chr2   330   340     B     .      -      4
#> 15  chr2   340   350     B     .      -      5
#> 

# Fixed window size with overlaps
bed_makewindows(x, genome, win_size = 10, step_size = 5)

#> # A tibble: 30 × 7
#>    chrom start   end  name score strand win_id
#>    <chr> <dbl> <dbl> <chr> <chr>  <chr>  <dbl>
#> 1   chr1   100   110     A     .      +      1
#> 2   chr1   105   115     A     .      +      2
#> 3   chr1   110   120     A     .      +      3
#> 4   chr1   115   125     A     .      +      4
#> 5   chr1   120   130     A     .      +      5
#> 6   chr1   125   135     A     .      +      6
#> 7   chr1   130   140     A     .      +      7
#> 8   chr1   135   145     A     .      +      8
#> 9   chr1   140   150     A     .      +      9
#> 10  chr1   145   155     A     .      +     10
#> # ... with 20 more rows
#> 

# reverse win_id
bed_makewindows(x, genome, win_size = 10, reverse = TRUE)

#> # A tibble: 15 × 7
#>    chrom start   end  name score strand win_id
#>    <chr> <dbl> <dbl> <chr> <chr>  <chr>  <dbl>
#> 1   chr1   100   110     A     .      +     10
#> 2   chr1   110   120     A     .      +      9
#> 3   chr1   120   130     A     .      +      8
#> 4   chr1   130   140     A     .      +      7
#> 5   chr1   140   150     A     .      +      6
#> 6   chr1   150   160     A     .      +      5
#> 7   chr1   160   170     A     .      +      4
#> 8   chr1   170   180     A     .      +      3
#> 9   chr1   180   190     A     .      +      2
#> 10  chr1   190   200     A     .      +      1
#> 11  chr2   300   310     B     .      -      5
#> 12  chr2   310   320     B     .      -      4
#> 13  chr2   320   330     B     .      -      3
#> 14  chr2   330   340     B     .      -      2
#> 15  chr2   340   350     B     .      -      1
#> 

# bedtools 'namenum'
wins <- bed_makewindows(x, genome, win_size = 10)
dplyr::mutate(wins, namenum = stringr::str_c(name, '_', win_id))

#> # A tibble: 15 × 8
#>    chrom start   end  name score strand win_id namenum
#>    <chr> <dbl> <dbl> <chr> <chr>  <chr>  <dbl>   <chr>
#> 1   chr1   100   110     A     .      +      1     A_1
#> 2   chr1   110   120     A     .      +      2     A_2
#> 3   chr1   120   130     A     .      +      3     A_3
#> 4   chr1   130   140     A     .      +      4     A_4
#> 5   chr1   140   150     A     .      +      5     A_5
#> 6   chr1   150   160     A     .      +      6     A_6
#> 7   chr1   160   170     A     .      +      7     A_7
#> 8   chr1   170   180     A     .      +      8     A_8
#> 9   chr1   180   190     A     .      +      9     A_9
#> 10  chr1   190   200     A     .      +     10    A_10
#> 11  chr2   300   310     B     .      -      1     B_1
#> 12  chr2   310   320     B     .      -      2     B_2
#> 13  chr2   320   330     B     .      -      3     B_3
#> 14  chr2   330   340     B     .      -      4     B_4
#> 15  chr2   340   350     B     .      -      5     B_5
#>