Projection test for query interval overlap.

Projection test for query interval overlap.

bed_projection(x, y, genome, by_chrom = FALSE)

Arguments

x
tbl_interval

y
tbl_interval

genome
tbl_genome

by_chrom
compute test per chromosome

Value

tbl_interval with the following columns:

chrom the name of chromosome tested if by_chrom = TRUE, otherwise has a value of 'whole_genome'

p.value p-value from a binomial test. p-values > 0.5 will be reported as 1 - p-value and lower_tail will be FALSE

obs_exp_ratio ratio of observed to expected overlap frequency

lower_tail a boolean column. TRUE indicates the observed overlaps is in the lower tail of the distribution (e.g., less overlap than expected); FALSE indicates that the observed overlaps are in the upper tail of the distribution (e.g., more overlap than expected)

Details

Interval statistics can be used in combination with group_by and do to calculate statistics for subsets of data. See vignette('interval-stats') for examples.

See also

http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002529

Other interval statistics: bed_absdist, bed_fisher, bed_jaccard, bed_reldist

Examples

genome <- trbl_genome(
 ~chrom, ~size,
 "chr1", 1e4,
 "chr2", 2e4,
 "chr3", 4e4
)

x <- trbl_interval(
 ~chrom, ~start, ~end,
 "chr1", 100,    200,
 "chr1", 250,    400,
 "chr1", 500,    600,
 "chr1", 1000,   2000,
 "chr2", 100,    200
)

y <- trbl_interval(
 ~chrom, ~start, ~end,
 "chr1", 150,    175,
 "chr1", 525,    575,
 "chr1", 1100,   1200,
 "chr1", 1400,   1600,
 "chr2", 200,    1500
)

bed_projection(x, y, genome)

#> # A tibble: 1 × 4
#>          chrom      p.value obs_exp_ratio lower_tail
#>          <chr>        <dbl>         <dbl>      <chr>
#> 1 whole_genome 4.641936e-05      10.74627      FALSE

bed_projection(x, y, genome, by_chrom = TRUE)

#> # A tibble: 2 × 4
#>   chrom      p.value obs_exp_ratio lower_tail
#>   <chr>        <dbl>         <dbl>      <chr>
#> 1  chr1 1.977539e-06            20      FALSE
#> 2  chr2 6.500000e-02             0      FALSE