read genome files (i.e., UCSC "chromSize" files)

Genome files contain chromosome name and size information. These sizes are used by downstream functions to identify computed intervals that have ends outside the defined size.

Select intervals bounded by a genome.

read_genome(filename)

bound_intervals(x, genome, trim = FALSE)

Arguments

filename

file containing chrom/contig names and sizes, one-pair-per-line, tab-delimited

x

a tbl of intervals

genome

a tbl of chrom sizes

trim

adjust coordinates for out-of-bounds intervals

Value

data_frame with colnames chrom and size, sorted by size

data_frame

See also

Other read-funcs: read_bed, read_vcf

Examples

genome <- valr_example('hg19.chrom.sizes.gz')
read_genome(genome)

#> # A tibble: 25 × 2
#>    chrom      size
#>    <chr>     <int>
#> 1   chr1 249250621
#> 2   chr2 243199373
#> 3   chr3 198022430
#> 4   chr4 191154276
#> 5   chr5 180915260
#> 6   chr6 171115067
#> 7   chr7 159138663
#> 8   chrX 155270560
#> 9   chr8 146364022
#> 10  chr9 141213431
#> # ... with 15 more rows
#> 

## Not run: ------------------------------------
# # can also read from a URL
# read_genome('https://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes')
## ---------------------------------------------

x <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", -100,   500,
 "chr1", 100,    1e9,
 "chr1", 500,    1000
)

genome <- read_genome(valr_example('hg19.chrom.sizes.gz'))

bound_intervals(x, genome)

#> # A tibble: 1 × 3
#>   chrom start   end
#>   <chr> <dbl> <dbl>
#> 1  chr1   500  1000
#> 

bound_intervals(x, genome, trim = TRUE)

#> # A tibble: 3 × 3
#>   chrom start       end
#>   <chr> <dbl>     <dbl>
#> 1  chr1     1       500
#> 2  chr1   100 249250621
#> 3  chr1   500      1000
#>