Compute absolute distances between intervals.

Computes the absolute distance between the midpoints of x intervals and the closest midpoints of y intervals.

bed_absdist(x, y, genome)

Arguments

x

tbl of intervals

y

tbl of intervals

genome

genome tbl

Value

data_frame with .absdist and .absdist_scaled columns.

Details

Absolute distances are scaled by the inter-reference gap for the chromosome as follows. For Q query points and R reference points on a chromosome, scale the distance for each query point i to the closest reference point by the inter-reference gap for each chromosome. If an x interval has no matching y chromosome, .absdist is NA. $$d_i(x,y) = min_k(|q_i - r_k|)\frac{R}{Length\ of\ chromosome}$$ Both absolute and scaled distances are reported as .absdist and .absdist_scaled.

Interval statistics can be used in combination with group_by and do to calculate statistics for subsets of data. See vignette('interval-stats') for examples.

See also

http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002529

Other interval statistics: bed_fisher, bed_jaccard, bed_projection, bed_reldist

Examples

x <- tibble::frame_data(
~chrom,   ~start,    ~end,
"chr1",    75,       125
  )

y <- tibble::frame_data(
  ~chrom,   ~start,    ~end,
  "chr1",    50,       100,
  "chr1",    100,       150
  )

genome <- tibble::frame_data(
  ~chrom, ~size,
  "chr1", 500,
  "chr2", 1000
)

bed_absdist(x, y, genome)

#> # A tibble: 1 × 5
#>   chrom start   end .absdist .absdist_scaled
#>   <chr> <dbl> <dbl>    <dbl>           <dbl>
#> 1  chr1    75   125       25             0.1