Calculate jaccard statistics on two sets of intervals.

Calculate jaccard statistics on two sets of intervals.

bed_jaccard(x, y)

Arguments

x

tbl of intervals

y

tbl of intervals

Value

data_frame with the following columns:

len_i length of the intersection

len_u length of the union

jaccard jaccard statistic

n_int number of intersecting intervals between x and y

Details

bed_jaccard() quantifies the extent of overlap between to sets of intervals. The Jaccard statistic takes values of [0,1] and is measured as:

$$ J(x,y) = \frac{\mid x \bigcap y \mid} {\mid x \bigcup y \mid} = \frac{\mid x \bigcap y \mid} {\mid x \mid + \mid y \mid - \mid x \bigcap y \mid} $$

See also

http://bedtools.readthedocs.org/en/latest/content/tools/jaccard.html

Other interval-stats: bed_absdist, bed_fisher, bed_reldist

Examples

x <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 10,     20,
  "chr1", 30,     40
)

y <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 15,     20
)

bed_jaccard(x, y)

#> # A tibble: 1 × 4
#>   len_i len_u jaccard     n
#>   <int> <dbl>   <dbl> <int>
#> 1     5    25    0.25     1
#>