Compute relative distances between intervals.

Compute relative distances between intervals.

bed_reldist(x, y, detail = FALSE)

Arguments

x
tbl_interval

y
tbl_interval

detail
report relative distances for each x interval.

Value

If detail = FALSE, a tbl_interval that summarizes calcuclated .reldist values with the following columns:

.reldist relative distance metric

.counts number of metric observations

.total total observations

.freq frequency of observation

If detail = TRUE, a new .reldist column reports the relative distance for each input x interval.

Details

Interval statistics can be used in combination with group_by and do to calculate statistics for subsets of data. See vignette('interval-stats') for examples.

See also

http://bedtools.readthedocs.io/en/latest/content/tools/reldist.html

Other interval statistics: bed_absdist, bed_fisher, bed_jaccard, bed_projection

Examples

x <- trbl_interval(
  ~chrom,   ~start,    ~end,
  "chr1",    75,       125
)

y <- trbl_interval(
  ~chrom,   ~start,    ~end,
  "chr1",    50,       100,
  "chr1",    100,       150
)

bed_reldist(x, y)

#> # A tibble: 1 × 4
#>   .reldist .counts .total .freq
#>      <dbl>   <int>  <int> <dbl>
#> 1      0.5       1      1     1

bed_reldist(x, y, detail = TRUE)

#> # A tibble: 1 × 4
#>   chrom start   end .reldist
#>   <chr> <dbl> <dbl>    <dbl>
#> 1  chr1    75   125      0.5